+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 
 ## [Unreleased]
+### Added
+- `?standards` command to display [xkcd 927](https://xkcd.com/927/).
 ### Changed
 - Python>=3.7 now required.